Progression on login.cpp
I was able to create a timer which counts down from 10 every time the user exceeds three login attempts. Once the timer is up, they can login again. The only issue is that the user is forced to login one more time then necessary to  be able to successfully login again.