Error checking for generating reports in admin
This was tricky to do because every single time a student record that fit the report criteria was found in a file, it would print out to the console and then the error message would print out if a match later on could not be found. I just ended up clearing the console for better readability if this occured.